+++ client/components/index.js
@@ -11,9 +11,9 @@ export {default as BooksByGenre} from './books-by-genre-view'
 export {Login, Signup} from './auth-form'
 export {default as LandingPage} from './landing-page'
 export {default as Cart} from './cart'
-export {default as Checkout} from './checkout'
-export {default as StripeCheckout} from './stripe-checkout'
-export {default as ReviewOrder} from './review-order'
+export {default as Checkout} from './payment/checkout'
+export {default as StripeCheckout} from './payment/stripe-checkout'
+export {default as ReviewOrder} from './payment/review-order'
 export {default as AllStaffs} from './all-staff'
 export {default as SingleStaff} from './single-staff'
 export {default as Footer} from './footer'
