+++ client/components/payment/accept-payment.js
