+++ client/components/payment/stripe-checkout.js
